proxy: move helper methods for muliplexed connections from go/cmd/vpnkit-forwarder to go/pkg/libproxy
